+++ zetaclient/orchestrator/orchestrator.go
@@ -12,9 +12,9 @@ import (
 	ethcommon "github.com/ethereum/go-ethereum/common"
 	"github.com/pkg/errors"
 	"github.com/rs/zerolog"
+	"github.com/samber/lo"
 
 	"github.com/zeta-chain/zetacore/pkg/bg"
-	"github.com/zeta-chain/zetacore/pkg/chains"
 	zetamath "github.com/zeta-chain/zetacore/pkg/math"
 	"github.com/zeta-chain/zetacore/x/crosschain/types"
 	observertypes "github.com/zeta-chain/zetacore/x/observer/types"
@@ -22,6 +22,7 @@ import (
 	btcobserver "github.com/zeta-chain/zetacore/zetaclient/chains/bitcoin/observer"
 	"github.com/zeta-chain/zetacore/zetaclient/chains/evm"
 	"github.com/zeta-chain/zetacore/zetaclient/chains/interfaces"
+	solanaobserver "github.com/zeta-chain/zetacore/zetaclient/chains/solana/observer"
 	zctx "github.com/zeta-chain/zetacore/zetaclient/context"
 	"github.com/zeta-chain/zetacore/zetaclient/metrics"
 	"github.com/zeta-chain/zetacore/zetaclient/outboundprocessor"
@@ -150,31 +151,42 @@ func (oc *Orchestrator) resolveSigner(app *zctx.AppContext, chainID int64) (inte
 		return nil, err
 	}
 
-	// noop for non-EVM chains
-	if !chains.IsEVMChain(chainID, app.GetAdditionalChains()) {
-		return signer, nil
-	}
+	chain, err := app.GetChain(chainID)
+	switch {
+	case err != nil:
+		return nil, err
+	case chain.IsZeta():
+		// should not happen
+		return nil, fmt.Errorf("unable to resolve signer for zeta chain %d", chainID)
+	case chain.IsEVM():
+		params := chain.Params()
 
-	evmParams, found := app.GetEVMChainParams(chainID)
-	if !found {
-		return signer, nil
-	}
+		// update zeta connector and ERC20 custody addresses
+		zetaConnectorAddress := ethcommon.HexToAddress(params.GetConnectorContractAddress())
+		if zetaConnectorAddress != signer.GetZetaConnectorAddress() {
+			signer.SetZetaConnectorAddress(zetaConnectorAddress)
+			oc.logger.Info().
+				Str("signer.connector_address", zetaConnectorAddress.String()).
+				Msgf("updated zeta connector address for chain %d", chainID)
+		}
 
-	// update zeta connector and ERC20 custody addresses
-	zetaConnectorAddress := ethcommon.HexToAddress(evmParams.GetConnectorContractAddress())
-	if zetaConnectorAddress != signer.GetZetaConnectorAddress() {
-		signer.SetZetaConnectorAddress(zetaConnectorAddress)
-		oc.logger.Info().
-			Str("signer.connector_address", zetaConnectorAddress.String()).
-			Msgf("updated zeta connector address for chain %d", chainID)
-	}
+		erc20CustodyAddress := ethcommon.HexToAddress(params.GetErc20CustodyContractAddress())
+		if erc20CustodyAddress != signer.GetERC20CustodyAddress() {
+			signer.SetERC20CustodyAddress(erc20CustodyAddress)
+			oc.logger.Info().
+				Str("signer.erc20_custody", erc20CustodyAddress.String()).
+				Msgf("updated zeta connector address for chain %d", chainID)
+		}
+	case chain.IsSolana():
+		params := chain.Params()
 
-	erc20CustodyAddress := ethcommon.HexToAddress(evmParams.GetErc20CustodyContractAddress())
-	if erc20CustodyAddress != signer.GetERC20CustodyAddress() {
-		signer.SetERC20CustodyAddress(erc20CustodyAddress)
-		oc.logger.Info().
-			Str("signer.erc20_custody", erc20CustodyAddress.String()).
-			Msgf("updated zeta connector address for chain %d", chainID)
+		// update solana gateway address
+		if params.GatewayAddress != signer.GetGatewayAddress() {
+			signer.SetGatewayAddress(params.GatewayAddress)
+			oc.logger.Info().
+				Str("signer.gateway_address", params.GatewayAddress).
+				Msgf("updated gateway address for chain %d", chainID)
+		}
 	}
 
 	return signer, nil
@@ -199,24 +211,26 @@ func (oc *Orchestrator) resolveObserver(app *zctx.AppContext, chainID int64) (in
 		return nil, err
 	}
 
+	chain, err := app.GetChain(chainID)
+	switch {
+	case err != nil:
+		return nil, errors.Wrapf(err, "unable to get chain %d", chainID)
+	case chain.IsZeta():
+		// should not happen
+		return nil, fmt.Errorf("unable to resolve observer for zeta chain %d", chainID)
+	}
+
 	// update chain observer chain parameters
-	curParams := observer.GetChainParams()
-	if chains.IsEVMChain(chainID, app.GetAdditionalChains()) {
-		evmParams, found := app.GetEVMChainParams(chainID)
-		if found && !observertypes.ChainParamsEqual(curParams, *evmParams) {
-			observer.SetChainParams(*evmParams)
-			oc.logger.Info().
-				Interface("observer.chain_params", *evmParams).
-				Msgf("updated chain params for EVM chainID %d", chainID)
-		}
-	} else if chains.IsBitcoinChain(chainID, app.GetAdditionalChains()) {
-		_, btcParams, found := app.GetBTCChainParams()
-		if found && !observertypes.ChainParamsEqual(curParams, *btcParams) {
-			observer.SetChainParams(*btcParams)
-			oc.logger.Info().
-				Interface("observer.chain_params", *btcParams).
-				Msgf("updated chain params for UTXO chainID %d", btcParams.ChainId)
-		}
+	var (
+		curParams   = observer.GetChainParams()
+		freshParams = chain.Params()
+	)
+
+	if !observertypes.ChainParamsEqual(curParams, *freshParams) {
+		observer.SetChainParams(*freshParams)
+		oc.logger.Info().
+			Interface("observer.chain_params", *freshParams).
+			Msgf("updated chain params for chainID %d", chainID)
 	}
 
 	return observer, nil
@@ -235,10 +249,10 @@ func (oc *Orchestrator) getObserver(chainID int64) (interfaces.ChainObserver, er
 }
 
 // GetPendingCctxsWithinRateLimit get pending cctxs across foreign chains within rate limit
-func (oc *Orchestrator) GetPendingCctxsWithinRateLimit(
-	ctx context.Context,
-	foreignChains []chains.Chain,
-) (map[int64][]*types.CrossChainTx, error) {
+func (oc *Orchestrator) GetPendingCctxsWithinRateLimit(ctx context.Context, chainIDs []int64) (
+	map[int64][]*types.CrossChainTx,
+	error,
+) {
 	// get rate limiter flags
 	rateLimitFlags, err := oc.zetacoreClient.GetRateLimiterFlags(ctx)
 	if err != nil {
@@ -251,10 +265,10 @@ func (oc *Orchestrator) GetPendingCctxsWithinRateLimit(
 	// fallback to non-rate-limited query if rate limiter is not usable
 	cctxsMap := make(map[int64][]*types.CrossChainTx)
 	if !rateLimiterUsable {
-		for _, chain := range foreignChains {
-			resp, _, err := oc.zetacoreClient.ListPendingCCTX(ctx, chain.ChainId)
+		for _, chainID := range chainIDs {
+			resp, _, err := oc.zetacoreClient.ListPendingCCTX(ctx, chainID)
 			if err == nil && resp != nil {
-				cctxsMap[chain.ChainId] = resp
+				cctxsMap[chainID] = resp
 			}
 		}
 		return cctxsMap, nil
@@ -334,49 +348,68 @@ func (oc *Orchestrator) runScheduler(ctx context.Context) error {
 					// set current hot key burn rate
 					metrics.HotKeyBurnRate.Set(float64(oc.ts.HotKeyBurnRate.GetBurnRate().Int64()))
 
-					// get supported external chains
-					externalChains := app.GetEnabledExternalChains()
+					// get chain ids without zeta chain
+					chainIDs := lo.FilterMap(app.ListChains(), func(c zctx.Chain, _ int) (int64, bool) {
+						return c.ID(), !c.IsZeta()
+					})
 
 					// query pending cctxs across all external chains within rate limit
-					cctxMap, err := oc.GetPendingCctxsWithinRateLimit(ctx, externalChains)
+					cctxMap, err := oc.GetPendingCctxsWithinRateLimit(ctx, chainIDs)
 					if err != nil {
 						oc.logger.Error().Err(err).Msgf("runScheduler: GetPendingCctxsWithinRatelimit failed")
 					}
 
 					// schedule keysign for pending cctxs on each chain
-					for _, c := range externalChains {
+					for _, chain := range app.ListChains() {
+						// skip zeta chain
+						if chain.IsZeta() {
+							continue
+						}
+
+						chainID := chain.ID()
+
 						// get cctxs from map and set pending transactions prometheus gauge
-						cctxList := cctxMap[c.ChainId]
-						metrics.PendingTxsPerChain.WithLabelValues(c.Name).Set(float64(len(cctxList)))
+						cctxList := cctxMap[chainID]
+
+						metrics.PendingTxsPerChain.
+							WithLabelValues(fmt.Sprintf("chain_%d", chainID)).
+							Set(float64(len(cctxList)))
+
 						if len(cctxList) == 0 {
 							continue
 						}
 
 						// update chain parameters for signer and chain observer
-						signer, err := oc.resolveSigner(app, c.ChainId)
+						signer, err := oc.resolveSigner(app, chainID)
 						if err != nil {
 							oc.logger.Error().Err(err).
-								Msgf("runScheduler: unable to resolve signer for chain %d", c.ChainId)
+								Msgf("runScheduler: unable to resolve signer for chain %d", chainID)
 							continue
 						}
-						ob, err := oc.resolveObserver(app, c.ChainId)
+
+						ob, err := oc.resolveObserver(app, chainID)
 						if err != nil {
 							oc.logger.Error().Err(err).
-								Msgf("runScheduler: resolveObserver failed for chain %d", c.ChainId)
+								Msgf("runScheduler: resolveObserver failed for chain %d", chainID)
 							continue
 						}
-						if !app.IsOutboundObservationEnabled(ob.GetChainParams()) {
+
+						if !app.IsOutboundObservationEnabled() {
 							continue
 						}
 
 						// #nosec G115 range is verified
 						zetaHeight := uint64(bn)
-						if chains.IsEVMChain(c.ChainId, app.GetAdditionalChains()) {
-							oc.ScheduleCctxEVM(ctx, zetaHeight, c.ChainId, cctxList, ob, signer)
-						} else if chains.IsBitcoinChain(c.ChainId, app.GetAdditionalChains()) {
-							oc.ScheduleCctxBTC(ctx, zetaHeight, c.ChainId, cctxList, ob, signer)
-						} else {
-							oc.logger.Error().Msgf("StartCctxScheduler: unsupported chain %d", c.ChainId)
+
+						switch {
+						case chain.IsEVM():
+							oc.ScheduleCctxEVM(ctx, zetaHeight, chainID, cctxList, ob, signer)
+						case chain.IsUTXO():
+							oc.ScheduleCctxBTC(ctx, zetaHeight, chainID, cctxList, ob, signer)
+						case chain.IsSolana():
+							oc.ScheduleCctxSolana(ctx, zetaHeight, chainID, cctxList, ob, signer)
+						default:
+							oc.logger.Error().Msgf("runScheduler: no scheduler found chain %d", chainID)
 							continue
 						}
 					}
@@ -430,17 +463,17 @@ func (oc *Orchestrator) ScheduleCctxEVM(
 			break
 		}
 
-		// try confirming the outbound
-		included, _, err := observer.IsOutboundProcessed(ctx, cctx, oc.logger.Logger)
+		// vote outbound if it's already confirmed
+		continueKeysign, err := observer.VoteOutboundIfConfirmed(ctx, cctx)
 		if err != nil {
 			oc.logger.Error().
 				Err(err).
-				Msgf("ScheduleCctxEVM: IsOutboundProcessed faild for chain %d nonce %d", chainID, nonce)
+				Msgf("ScheduleCctxEVM: VoteOutboundIfConfirmed failed for chain %d nonce %d", chainID, nonce)
 			continue
 		}
-		if included {
+		if !continueKeysign {
 			oc.logger.Info().
-				Msgf("ScheduleCctxEVM: outbound %s already included; do not schedule keysign", outboundID)
+				Msgf("ScheduleCctxEVM: outbound %s already processed; do not schedule keysign", outboundID)
 			continue
 		}
 
@@ -470,7 +503,7 @@ func (oc *Orchestrator) ScheduleCctxEVM(
 			!oc.outboundProc.IsOutboundActive(outboundID) {
 			oc.outboundProc.StartTryProcess(outboundID)
 			oc.logger.Debug().
-				Msgf("ScheduleCctxEVM: sign outbound %s with value %d\n", outboundID, cctx.GetCurrentOutboundParam().Amount)
+				Msgf("ScheduleCctxEVM: sign outbound %s with value %d", outboundID, cctx.GetCurrentOutboundParam().Amount)
 			go signer.TryProcessOutbound(
 				ctx,
 				cctx,
@@ -522,16 +555,16 @@ func (oc *Orchestrator) ScheduleCctxBTC(
 			continue
 		}
 		// try confirming the outbound
-		included, confirmed, err := btcObserver.IsOutboundProcessed(ctx, cctx, oc.logger.Logger)
+		continueKeysign, err := btcObserver.VoteOutboundIfConfirmed(ctx, cctx)
 		if err != nil {
 			oc.logger.Error().
 				Err(err).
-				Msgf("ScheduleCctxBTC: IsOutboundProcessed faild for chain %d nonce %d", chainID, nonce)
+				Msgf("ScheduleCctxBTC: VoteOutboundIfConfirmed failed for chain %d nonce %d", chainID, nonce)
 			continue
 		}
-		if included || confirmed {
+		if !continueKeysign {
 			oc.logger.Info().
-				Msgf("ScheduleCctxBTC: outbound %s already included; do not schedule keysign", outboundID)
+				Msgf("ScheduleCctxBTC: outbound %s already processed; do not schedule keysign", outboundID)
 			continue
 		}
 
@@ -547,10 +580,70 @@ func (oc *Orchestrator) ScheduleCctxBTC(
 				Msgf("ScheduleCctxBTC: lookahead reached, signing %d, earliest pending %d", nonce, cctxList[0].GetCurrentOutboundParam().TssNonce)
 			break
 		}
-		// try confirming the outbound or scheduling a keysign
+		// schedule a TSS keysign
 		if nonce%interval == zetaHeight%interval && !oc.outboundProc.IsOutboundActive(outboundID) {
 			oc.outboundProc.StartTryProcess(outboundID)
-			oc.logger.Debug().Msgf("ScheduleCctxBTC: sign outbound %s with value %d\n", outboundID, params.Amount)
+			oc.logger.Debug().Msgf("ScheduleCctxBTC: sign outbound %s with value %d", outboundID, params.Amount)
+			go signer.TryProcessOutbound(
+				ctx,
+				cctx,
+				oc.outboundProc,
+				outboundID,
+				observer,
+				oc.zetacoreClient,
+				zetaHeight,
+			)
+		}
+	}
+}
+
+// ScheduleCctxSolana schedules solana outbound keysign on each ZetaChain block (the ticker)
+func (oc *Orchestrator) ScheduleCctxSolana(
+	ctx context.Context,
+	zetaHeight uint64,
+	chainID int64,
+	cctxList []*types.CrossChainTx,
+	observer interfaces.ChainObserver,
+	signer interfaces.ChainSigner,
+) {
+	solObserver, ok := observer.(*solanaobserver.Observer)
+	if !ok { // should never happen
+		oc.logger.Error().Msgf("ScheduleCctxSolana: chain observer is not a solana observer")
+		return
+	}
+	// #nosec G701 positive
+	interval := uint64(observer.GetChainParams().OutboundScheduleInterval)
+
+	// schedule keysign for each pending cctx
+	for _, cctx := range cctxList {
+		params := cctx.GetCurrentOutboundParam()
+		nonce := params.TssNonce
+		outboundID := outboundprocessor.ToOutboundID(cctx.Index, params.ReceiverChainId, nonce)
+
+		if params.ReceiverChainId != chainID {
+			oc.logger.Error().
+				Msgf("ScheduleCctxSolana: outbound %s chainid mismatch: want %d, got %d", outboundID, chainID, params.ReceiverChainId)
+			continue
+		}
+
+		// vote outbound if it's already confirmed
+		continueKeysign, err := solObserver.VoteOutboundIfConfirmed(ctx, cctx)
+		if err != nil {
+			oc.logger.Error().
+				Err(err).
+				Msgf("ScheduleCctxSolana: VoteOutboundIfConfirmed failed for chain %d nonce %d", chainID, nonce)
+			continue
+		}
+		if !continueKeysign {
+			oc.logger.Info().
+				Msgf("ScheduleCctxSolana: outbound %s already processed; do not schedule keysign", outboundID)
+			continue
+		}
+
+		// schedule a TSS keysign
+		if nonce%interval == zetaHeight%interval && !oc.outboundProc.IsOutboundActive(outboundID) {
+			oc.outboundProc.StartTryProcess(outboundID)
+			oc.logger.Debug().Msgf("ScheduleCctxSolana: sign outbound %s with value %d", outboundID, params.Amount)
 			go signer.TryProcessOutbound(
 				ctx,
 				cctx,