+++ changelog.md
@@ -3,11 +3,16 @@
 ## Unreleased
 
 ### Features
+
+* [2578](https://github.com/zeta-chain/node/pull/2578) - Add Gateway address in protocol contract list
 * [2538](https://github.com/zeta-chain/node/pull/2538) - add background worker routines to shutdown zetaclientd when needed for tss migration
+
+## v19.0.0
 
 ### Breaking Changes
 
 * [2460](https://github.com/zeta-chain/node/pull/2460) - Upgrade to go 1.22. This required us to temporarily remove the QUIC backend from [go-libp2p](https://github.com/libp2p/go-libp2p). If you are a zetaclient operator and have configured quic peers, you need to switch to tcp peers.
+* [List of the other breaking changes can be found in this document](docs/releases/v19_breaking_changes.md)
 
 ### Features
 
@@ -39,6 +44,9 @@
 * [2518](https://github.com/zeta-chain/node/pull/2518) - add support for Solana address in zetacore
 * [2483](https://github.com/zeta-chain/node/pull/2483) - add priorityFee (gasTipCap) gas to the state
 * [2567](https://github.com/zeta-chain/node/pull/2567) - add sign latency metric to zetaclient (zetaclient_sign_latency)
+* [2524](https://github.com/zeta-chain/node/pull/2524) - add inscription envelop parsing 
+* [2560](https://github.com/zeta-chain/node/pull/2560) - add support for Solana SOL token withdraw
+* [2533](https://github.com/zeta-chain/node/pull/2533) - parse memo from both OP_RETURN and inscription
 
 ### Refactor
 
@@ -71,7 +79,9 @@
 * [2515](https://github.com/zeta-chain/node/pull/2515) - replace chainName by chainID for ChainNonces indexing
 * [2541](https://github.com/zeta-chain/node/pull/2541) - deprecate ChainName field in Chain object
 * [2542](https://github.com/zeta-chain/node/pull/2542) - adjust permissions to be more restrictive
+* [2572](https://github.com/zeta-chain/node/pull/2572) - turn off IBC modules
 * [2556](https://github.com/zeta-chain/node/pull/2556) - refactor migrator length check to use consensus type
+* [2568](https://github.com/zeta-chain/node/pull/2568) - improve AppContext by converging chains, chainParams, enabledChains, and additionalChains into a single zctx.Chain
 
 ### Tests